+++ imcger/recenttopicsng/language/de_x_sie/info_acp_rtng.php
@@ -41,20 +41,26 @@ if (empty($lang) || !is_array($lang))
 $lang = array_merge($lang, [
 	// Language pack author
 	'RTNG_LANG_DESC'				=> 'Deutsch (Sie)',
-	'RTNG_LANG_EXT_VER' 			=> '1.0.0',
+	'RTNG_LANG_VER' 				=> '1.1.0',
 	'RTNG_LANG_AUTHOR' 				=> 'IMC-GER / LukeWCS',
 
-	//forum acp
+	// ACP forums
 	'RTNG_FORUMS'					=> 'In „Recent Topics NG“ anzeigen',
 	'RTNG_FORUMS_EXPLAIN'			=> 'Aktiviere dieses Kontrollkästchen, um Themen dieses Forum in den Aktuelle Themen anzuzeigen.',
 
-	//acp title
-	'RTNG_NAME'						=> 'Recent Topics NG', // Please do not translate the name of the extension
-	'RTNG_DESIG'					=> 'Aktuelle Themen',
-	'RTNG_EXPLAIN'					=> 'Auf dieser Seite können Sie die Einstellungen der Erweiterung „Recent Topics NG“ anpassen.<br><br>Spezifische Foren können eingeschlossen oder ausgeschlossen werden.<br>Überprüfe auch die Benutzerberechtigungen, welche Benutzern erlauben, einige der Parameter für sich selbst zu verändern. Diese haben dann Vorrang vor den Einstellungen des Admin-Panels.',
+	// ACP nav
+	'RTNG_NAME'						=> 'Aktuelle Themen',
 	'RTNG_CONFIG'					=> 'Einstellungen',
 
-	//allgemeine Einstellungen
+	// ACP module
+	'RTNG_EXPLAIN'					=> 'Auf dieser Seite können Sie die Einstellungen der Erweiterung <strong>„%s“</strong> anpassen.<br><br>Spezifische Foren können eingeschlossen oder ausgeschlossen werden.<br>Überprüfe auch die Benutzerberechtigungen, welche Benutzern erlauben, einige der Parameter für sich selbst zu verändern. Diese haben dann Vorrang vor den Einstellungen des Admin-Panels.',
+
+	// ACP load
+	'RTNG_LOAD_OPTIONS'					=> '„Recent Topics NG“ Optionen',
+	'RTNG_LOAD_FIRST_UNRD_POST'			=> 'Ermöglicht den Zugriff auf den ersten ungelesenen Beitrag',
+	'RTNG_LOAD_FIRST_UNRD_POST_EXPLAIN' => 'Wenn diese Option aktiviert ist, werden die Daten des ersten ungelesenen Beitrags ausgelesen und zur weiteren Verarbeitung bereitgestellt.<br>„<u><a href="#load_db_lastread">Serverseitige Gelesen-Markierung aktivieren</a></u>“ muss aktiviert sein.',
+
+	// Global settings
 	'RTNG_GLOBAL_SETTINGS'			=> 'Globale Einstellungen',
 	'RTNG_INDEX_DISPLAY_EXP'		=> 'Anzeigen auf der Index-Seite.',
 	'RTNG_ALL_TOPICS'				=> 'Alle Seiten anzeigen',
@@ -71,8 +77,9 @@ $lang = array_merge($lang, [
 	'RTNG_SIMPLE_PAGE_QTY'			=> 'Anzahl der Seiten in der vereinfachten Anzeige',
 	'RTNG_SIMPLE_PAGE_QTY_EXP'		=> 'Damit können Sie für die vereinfachte Anzeige festlegen, wie viele Listen-Seiten maximal angezeigt werden sollen.',
 
-	//Benutzereinstellungen
+	// User Overridable settings. these apply for anon users and can be overridden by UCP
 	'RTNG_OVERRIDABLE'				=> 'Einstellungen, die im Benutzerkontrollzentrum geändert werden können',
+	'RTNG_OVERRIDABLE_EXPLAIN'		=> 'Damit der Benutzer diese Einstellungen im Benutzerkontrollzentrum verändern kann, muss ihm in der Rechteverwaltung das entsprechende Benutzerrecht zugewiesen werden. Hat er dieses Recht nicht, werden für ihn diese Standardeinstellungen angewendet. Diese Werte werden auch für neue Benutzer und Gäste gesetzt.<br><br>Damit der Titel, Autor und das Beitragsdatum des ersten ungelesenen Beitrags als Anzeige im Thementitel ausgewählt werden kann, muss diese Option in den Einstellungen der <u><a href="%s">Serverlast</a></u> aktiviert werden.',
 	'RTNG_ENABLE'					=> 'Aktuelle Themen anzeigen',
 	'RTNG_LOCATION'					=> 'Anzeigeort',
 	'RTNG_LOCATION_EXP'				=> 'Wähle den Anzeigeort der aktuellen Themen.',