+++ imcger/recenttopicsng/language/en/rtng_common.php
@@ -39,11 +39,11 @@ if (empty($lang) || !is_array($lang))
 // ‚ ‘ ’ « » “ ” … „ “
 //
 $lang = array_merge($lang, [
-	'RTNG_NAME'			=> 'Recent Topics NG',
-	'RTNG_DESIG'		=> 'Recent Topics',
-	'RTNG_NO_TOPICS'	=> 'There are no new topics to display.',
-	'RTNG_TOPICS_COUNT' => '%d Topics',
+	'RTNG_NO_TOPICS'		=> 'There are no new topics to display.',
+	'RTNG_TOPICS_COUNT'		=> '%d Topics',
 
-	'RTNG_READ_SEPARATE' => 'Reading “Recent Topics”',
-	'RTNG_READ_SIMPLE'	 => 'Reading “Recent Topics” (simplified page)',
+	'RTNG_TITLE'			=> 'Recent Topics',
+	'RTNG_UNREAD_TITLE'		=> 'Unread Topics',
+	'RTNG_READ_SEPARATE'	=> 'Reading “%s”',
+	'RTNG_READ_SIMPLE'		=> 'Reading “%s” (simplified page)',
 ]);